backend/requirements.txt
# ============================================================================
# HireSight Unified API Requirements
# ============================================================================
# Pinned versions for reproducible builds
# Last updated: 2025-10-29
# Python 3.8+ required

# ============================================================================
# Flask web framework and extensions
# ============================================================================
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Rate limiting for API security
flask-limiter==3.5.0

# ============================================================================
# GitHub analysis dependencies
# ============================================================================
PyGithub==2.1.1
requests==2.31.0

# ============================================================================
# Data processing and analysis
# ============================================================================
# Updated for Python 3.13 compatibility
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2

# ============================================================================
# Text processing
# ============================================================================
textblob==0.17.1
beautifulsoup4==4.12.2
html2text==2020.1.16

# ============================================================================
# Date/time utilities
# ============================================================================
python-dateutil==2.8.2

# ============================================================================
# Configuration file support
# ============================================================================
PyYAML==6.0.1
toml==0.10.2

# ============================================================================
# Document processing (resume parsing)
# ============================================================================
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2

# ============================================================================
# Progress tracking
# ============================================================================
tqdm==4.66.1

# ============================================================================
# Production deployment
# ============================================================================
gunicorn==21.2.0
python-dotenv==1.0.0

# ============================================================================
# Development and testing (optional)
# ============================================================================
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.12.1
# flake8==6.1.0
# mypy==1.7.1
